Move Authorization out of ConfigureServices into an extension.
