+++ src/Employer.Web/Startup.ConfigureServices.cs
@@ -18,16 +18,13 @@ namespace Esfa.Recruit.Employer.Web
         private IConfiguration _configuration { get; }
         private IHostingEnvironment _hostingEnvironment { get; }
         private AuthenticationConfiguration _authConfig { get; }
-        private readonly IGetAssociatedEmployerAccountsService _accountsSvc;
         
         public Startup(IConfiguration config, IHostingEnvironment env)
         {
             _configuration = config;
             _hostingEnvironment = env;
             _authConfig = _configuration.GetSection("Authentication").Get<AuthenticationConfiguration>();
-
-            _accountsSvc = new GetAssociatedEmployerAccountsService();
-
+            
             if (env.IsDevelopment()  && _authConfig.IsEnabledForDev == false)
             {
                 _isAuthEnabled = false;
@@ -38,6 +35,9 @@ namespace Esfa.Recruit.Employer.Web
         public void ConfigureServices(IServiceCollection services)
         {
             services.AddIoC(_configuration);
+
+            //A service provider for resolving services configured in IoC
+            var sp = services.BuildServiceProvider();
 
             // Routing has to come before adding Mvc
             services.AddRouting(opt =>
@@ -52,47 +52,10 @@ namespace Esfa.Recruit.Employer.Web
 
             if (_isAuthEnabled)
             {
-                JwtSecurityTokenHandler.DefaultInboundClaimTypeMap.Clear();
-
-                services.AddAuthentication(options =>
-                {
-                    options.DefaultScheme = "Cookies";
-                    options.DefaultChallengeScheme = "oidc";
-                })
-                .AddCookie("Cookies", options =>
-                {
-                    options.AccessDeniedPath = "/Error/403";
-                })
-                .AddOpenIdConnect("oidc", options =>
-                {
-                    options.SignInScheme = "Cookies";
-
-                    options.Authority = _authConfig.Authority;
-                    options.MetadataAddress = _authConfig.MetaDataAddress;
-                    options.RequireHttpsMetadata = false;
-                    options.ResponseType = "code";
-                    options.ClientId = _authConfig.ClientId;
-                    options.ClientSecret = _authConfig.ClientSecret;
-                    options.Scope.Add("profile");
-
-                    options.Events.OnTokenValidated = PopulateAccountsClaim;
-                });
-
+                services.AddAuthenticationService(_authConfig, sp.GetService<IGetAssociatedEmployerAccountsService>());
                 services.AddAuthorizationService();
             }
         }
-
-        private Task PopulateAccountsClaim(TokenValidatedContext ctx)
-        {
-            var userId = ctx.Principal.Claims.First(c => c.Type.Equals(EmployerRecruitClaims.IdamsUserIdClaimTypeIdentifier)).Value;
-            var accounts = _accountsSvc.GetAssociatedAccounts(userId);
-
-            var accountsConcatenated = string.Join(",", accounts);
-            var associatedAccountClaim = new Claim(EmployerRecruitClaims.AccountsClaimsTypeIdentifier, accountsConcatenated, ClaimValueTypes.String);
-
-            ctx.Principal.Identities.First().AddClaim(associatedAccountClaim);
-
-            return Task.CompletedTask;
-        }
+        
     }
 }
